+++ RELEASE_NOTES.md
@@ -3,7 +3,6 @@
 **✨ Cải tiến:**
 - Cải thiện UI
 - Giao diện Classic — dành cho người dùng thích sự đơn giản
-- 
 ![NexusKey Classic](docs/images/nexus-classic.png)
 
 **🛠 Sửa lỗi:**
@@ -13,7 +12,10 @@
 - Thay thế chức năng nhấn Ctrl để tắt kiểm tra chính tả thành danh sách loại trừ thủ công
 - Sửa lỗi chuyển mã không hoạt động
 
-Lưu ý: Sức người có hạn nên phiên bản classic chắc chắn vẫn có lỗi, vui lòng đóng góp trong mục https://github.com/phatMT97/NexusKey/issues mình sẽ giải quyết từ từ :v
+Lưu ý: 
+1. Sức người có hạn không thể test hết toàn bộ nên phiên bản classic chắc chắn vẫn có lỗi(à thật ra bản nào cũng sẽ có :V), vui lòng đóng góp trong mục https://github.com/phatMT97/NexusKey/issues mình sẽ giải quyết từ từ :v
+2. Nếu bạn chuyển từ phiên bản modern sang classic hãy tắt chức năng khởi động cùng windows đi để tránh xung đột
+3. Đặt vào cùng chỗ với file config.toml, mọi config sẽ được load tự động không cần config thủ công, nếu không -> bug :V vui lòng báo cáo
 
 ---
 Cảm ơn bạn đã chọn NexusKey! Mọi góp ý của bạn đều giúp bộ gõ ngày càng hoàn thiện hơn.